# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CMS-Auto API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/complaints/{id}/attachments": {
      "post": {
        "tags": [
          "Attachments"
        ],
        "operationId": "complaintAttachmentUpload",
        "summary": "Upload a staff complaint attachment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachmentUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attachment uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid attachment request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Role, branch scope, or CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Complaint not found (COMPLAINT_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Attachments"
        ],
        "operationId": "complaintAttachmentList",
        "summary": "List staff complaint attachments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Staff complaint attachments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Role or branch scope denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Complaint not found (COMPLAINT_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/complaints/{id}/attachments/{attachmentId}/download": {
      "get": {
        "tags": [
          "Attachments"
        ],
        "operationId": "complaintAttachmentDownloadPrepare",
        "summary": "Prepare a staff attachment download",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attachmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attachment download prepared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentDownloadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Role or branch scope denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Complaint or attachment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/portal/attachments": {
      "post": {
        "tags": [
          "Portal"
        ],
        "operationId": "portalAttachmentUpload",
        "summary": "Upload a verified portal attachment",
        "parameters": [
          {
            "name": "x-portal-session",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachmentUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attachment uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, expired session, terminal complaint, or invalid attachment (PORTAL_VERIFICATION_FAILED/VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/communication-groups": {
      "get": {
        "tags": ["Communication groups"],
        "operationId": "communicationGroupsList",
        "summary": "List server-scoped personal and shared communication groups",
        "responses": { "200": { "description": "Communication groups", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommunicationGroupsResponse" } } } }, "401": { "description": "Missing or invalid session" } }
      },
      "post": {
        "tags": ["Communication groups"],
        "operationId": "communicationGroupsCreate",
        "summary": "Create a personal group or an Admin-managed shared group",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommunicationGroupWrite" } } } },
        "responses": { "201": { "description": "Communication group created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommunicationGroup" } } } }, "400": { "description": "Invalid request" }, "403": { "description": "Role or staff scope denied" } }
      }
    },
    "/communication-groups/{id}": {
      "patch": { "tags": ["Communication groups"], "operationId": "communicationGroupsUpdate", "summary": "Update an owned personal group or Admin-managed shared group", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommunicationGroupWrite" } } } }, "responses": { "200": { "description": "Communication group updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommunicationGroup" } } } }, "403": { "description": "Ownership or role denied" } } },
      "delete": { "tags": ["Communication groups"], "operationId": "communicationGroupsDeactivate", "summary": "Deactivate a communication group without deleting history", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Group deactivated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkResponse" } } } }, "403": { "description": "Ownership or role denied" } } }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message",
          "correlationId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "actualRecipientCount": {
            "type": "integer",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "FieldError": {
        "type": "object",
        "required": [
          "field",
          "code",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "AttachmentUploadRequest": {
        "type": "object",
        "required": [
          "fileName",
          "contentType",
          "sizeBytes",
          "contentBase64"
        ],
        "properties": {
          "fileName": {
            "type": "string",
            "minLength": 1
          },
          "contentType": {
            "type": "string",
            "minLength": 1
          },
          "sizeBytes": {
            "type": "integer",
            "minimum": 1
          },
          "contentBase64": {
            "type": "string",
            "minLength": 1
          },
          "customerVisible": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "Attachment": {
        "type": "object",
        "required": [
          "id",
          "complaintId",
          "fileName",
          "contentType",
          "sizeBytes",
          "scanStatus",
          "customerVisible"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "complaintId": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "minimum": 1
          },
          "scanStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "CLEAN",
              "REJECTED"
            ]
          },
          "customerVisible": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AttachmentUploadResponse": {
        "type": "object",
        "required": [
          "attachment"
        ],
        "properties": {
          "attachment": {
            "$ref": "#/components/schemas/Attachment"
          }
        },
        "additionalProperties": false
      },
      "AttachmentDownloadResponse": {
        "type": "object",
        "required": [
          "download"
        ],
        "properties": {
          "download": {
            "type": "object",
            "required": [
              "attachmentId",
              "token",
              "expiresAt"
            ],
            "properties": {
              "attachmentId": {
                "type": "string"
              },
              "token": {
                "type": "string",
                "minLength": 1
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
...
      "OkResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "AttachmentListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          }
        },
        "additionalProperties": false
      },
...
      "CommunicationGroupMember": {
        "type": "object",
        "required": ["userId", "displayName", "displayNameAr"],
        "properties": { "userId": { "type": "string" }, "displayName": { "type": "string" }, "displayNameAr": { "type": "string" } },
        "additionalProperties": false
      },
      "CommunicationGroup": {
        "type": "object",
        "required": ["id", "name", "visibility", "ownerId", "members", "createdAt", "updatedAt"],
        "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "visibility": { "type": "string", "enum": ["PERSONAL", "SHARED"] }, "ownerId": { "type": "string" }, "members": { "type": "array", "items": { "$ref": "#/components/schemas/CommunicationGroupMember" } }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" } },
        "additionalProperties": false
      },
      "CommunicationGroupWrite": {
        "type": "object",
        "required": ["name", "visibility", "memberUserIds"],
        "properties": { "name": { "type": "string", "minLength": 1, "maxLength": 80 }, "visibility": { "type": "string", "enum": ["PERSONAL", "SHARED"] }, "memberUserIds": { "type": "array", "maxItems": 100, "items": { "type": "string", "minLength": 1 } } },
        "additionalProperties": false
      },
      "CommunicationGroupsResponse": {
        "type": "object",
        "required": ["items", "eligibleMembers", "canManageShared"],
        "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/CommunicationGroup" } }, "eligibleMembers": { "type": "array", "items": { "$ref": "#/components/schemas/CommunicationGroupMember" } }, "canManageShared": { "type": "boolean" } },
        "additionalProperties": false
      }
    }
  }
}